Fix position delta field type

+++ Sources/GPSEntities/Adhoc/ConstructorStanding.swift
@@ -19,7 +19,7 @@ public struct ConstructorStanding: Codable {
         previousPoints: Double,
         previousMainColor: String,
         previousPosition: Int,
-        positionDelta: Double,
+        positionDelta: Int,
         raceName: String,
         raceFlag: String,
         raceDate: Date,
@@ -54,7 +54,7 @@ public struct ConstructorStanding: Codable {
     public let previousPoints: Double
     public let previousMainColor: String
     public let previousPosition: Int
-    public let positionDelta: Double
+    public let positionDelta: Int
     public let raceName: String
     public let raceFlag: String
     public let raceDate: Date
